feat(routing): use React router
