+++ client/src/main.tsx
@@ -1,10 +1,12 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import App from './App.tsx'
+import { App } from './App'
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+
+import { BrowserRouter } from "react-router-dom";
 
 import { AuthProvider } from "./contexts/";
 
@@ -14,8 +16,10 @@ createRoot(document.getElementById('root')!).render(
     <StrictMode>
         <QueryClientProvider client={queryClient}>
             <AuthProvider>
-                <App />
-                <ReactQueryDevtools initialIsOpen={false} />
+                <BrowserRouter>
+                    <App />
+                    <ReactQueryDevtools initialIsOpen={false} />
+                </BrowserRouter>
             </AuthProvider>
         </QueryClientProvider>
     </StrictMode>,
