+++ client/src/pages/HomePage.tsx
@@ -1,0 +1,7 @@
+export function HomePage() {
+    return (
+        <>
+            <h2>Welcome!</h2>
+        </>
+    )
+}
