+++ client/src/pages/LoginPage.tsx
@@ -1,0 +1,10 @@
+import { LoginForm } from "../components";
+
+export function LoginPage() {
+    return (
+        <>
+            <h2>Login</h2>
+            <LoginForm />
+        </>
+    )
+}
