+++ client/src/pages/UsersPage.tsx
@@ -1,0 +1,10 @@
+import { UserList } from "../components";
+
+export function UsersPage() {
+    return (
+        <>
+            <h2>Users</h2>
+            <UserList />
+        </>
+    )
+}
