+++ client/src/pages/index.ts
@@ -1,0 +1,3 @@
+export { HomePage } from "./HomePage";
+export { UsersPage } from "./UsersPage";
+export { LoginPage } from "./LoginPage";
